Planet Labels and Production Rates
Not sure if Roy implemented this yet but I added a public int to the
PlayerPlanet script called productionRate. This is multiplied by delta
time to produce ships.

I also added a GUIText to PlayerPlanet.cs to show the number of ships at
the planet. The ObjectLabel prefab is made a child of the a planet. The
planet is put in the Target field on the ObjectLabel's ObjectLabel
script.