Changing Planet Ownership
Planet ownership depends on the tag. I added OnTriggerEnter() to
PlayerShip.cs. The ship ignores ship collisions and checks the owner of
the planet. If the tag of the planet is the same as the owner string on
the ship, it increments the ship count. If its not the player's planet,
it decrements the ship count on the planet.

I added IncrementShips() and DecrementShips(string owner) to
PlayerPlanet.cs. The ship tells the planet to increment or decrement
when it arrives.

I added the public string owner to PlayerShip.cs. The only problem now
is getting "owner" assigned when the ship is instantiated.