Smarter AI
The AI now behaves a little more intelligently. It keeps track of its
total number of ships. Whenever a non-AI planet has fewer than the AI's
total number of ships, it sends all its ships to capture that planet.

Because we currently send all ships from a planet, this makes it easier
for the AI to steal planets.

Distance (and the time taken to travel) is not currently factored into
the AI's decision-making process.