Added an Instructions section and fixed some bugs
The bugs I fixed were:
1. Issues caused by a selected planet being taken over by the enemy.
2. Premature "You Lost" messages that came from not checking for player
ships.

Other changes:
Simplified the win/lose code: it no longer goes through a list of all
the planets to determine if you won or lost. Instead, it checks to see
if there are any player planets or enemy planets, and then if there are
player ships or enemy ships, and determines if you won or lost based on
that.