2 New Levels and Discovered Bugs
Added levels 3 and 4. Updated the build settings.

I found 2 bugs:
1) The production rate on planets change when they are captured. A
neutral planet can be set to a rate of 1 and changes to 3 when a player
captures it or a 2 when the enemy captures it.

2) Selection bug still exists. This may be due to the new prefabs I
created to account for enemy/neutral/player sounds.

I'll work on fixing these bugs and slowing down the AI before making any
more levels.