2 Levels and AI Changes
Added levels 8 and 9. Updated build settings accordingly. Changed the
neutral planet prefab so that it has a default production of 1.

Made some changes to the AI in GameController. I'll continue to tweak it
a little to simulate intelligence. However, I'll need to rework the AI
and make it run as a thread in order to get AI controlled planets to
behave more independently.