Added buttons to select all planets, set number of ships sent. Also various bug fixes.
Set up two new buttons: One for selecting all player planets, the other
for setting the number of ships to be sent, in percentage (TODO: put a
percentage sign). Also cleaned up the code in general a bit to make it a
little more readable, fixed the selection bug (Mostly) by:
1. changing Input.GetButton("Fire1") to Input.GetMouseButtonDown(0),
which makes it so it only registers one click.
2. making it remove the target as soon as it is done sending ships, so
it doesn't get stuck on a target.
Made a Target method in GamController which is used to select all the
planets.
I also made the "Canvas" a prefab, so that any alterations or additions
to the GUI will automatically be changed in all the levels.
I commented out a few lines of code that didn't seem necessary, but feel
free to uncomment them if you think we need them.