Added new planet features and fixed bugs
I made the following changes:
1. Set the production rate to be based on planet size: it now has an
inverse relationship, currently set to 1 / planet scale.
2. Set the ship capacity to be based on planet size: currently set to 25
* planet scale.
3. Made it so the larger the planet, the more ships it takes to conquer,
and the smaller the planet the easier to conquer.  This is also based on
planet scale: it subtracts the ships by 1 / planet scale, rather than
just subtracting by 1.
4. Fixed a bug in the send amount button.  I originally had it send the
percentage to only player planets, but that caused an issue if you took
over a planet and sent ships from it before you pushed the button again.
Since the button only affects planets with the player status, I simply
changed it to send the instructions to all planets.