Added a Pause menu, and set up difficulty
Changed the following:
1. Added a Pause menu, which can be accessed by pressing Spacebar or
Enter. The Pause menu has 4 buttons:
a. Restart: Restarts the level
b. Main Menu: Takes you to the top menu, saving the level you're in.
c. Resume: Resumes play.
2. Added a "Resume" button to the start page.  If you have played a
level and pressed "Main Menu", you can go back to the level you stopped
on.  I might change this so it just automatically sets the farthest
level you're on.
3. Added a Difficulty placeholder.  You can set the difficulty on the
Start menu and the Pause menu.  This changes the "difficulty" variable
in PlayerPrefs to either 1, 2 or 3, 1 being easy, 2 being medium, 3
being hard.  At the moment this doesn't do anything, but it should be
easy to add it to the code to change the difficulty.

Both the Resume and the Difficulty set variables in the PlayerPrefs
class, which is stored on the computer and will remain through multiple
plays, even if the window is closed.
TODO: add a reset button to the start page so you can reset the
PlayerPrefs class. For now just pressing "Start" rather than "Resume"
has the same effect.