Added a rudimentary difficulty system and improved UI
Added a simple difficulty system, just makes the AI faster if on a
harder difficulty.  I changed the buttons so they were no longer that
nasty green/red combo, now they are white and red, and all the menu
buttons have the same size font and same color.  The send amount and
select all buttons are the same, I figured Arial was easier to read at
that size and with numbers.